produce graph underlying Simba PBG embedding for an sc-RNA-seq SCE (or h5ad)

produce graph underlying Simba PBG embedding for an sc-RNA-seq SCE (or h5ad)

gen_graph_sce(h5adpath, simconf = simba_config(gen_graph_copy = FALSE))

Arguments

h5adpath

character(1) path to h5ad

simconf

output of `simba_config`

Value

an instance of S3 class `simba_scrna`, a list with elements gdf (a data.frame of source, relation, destination, only produced if gen_graph_copy is TRUE in simba_config), work_contents (full pathnames of pbg outputs), workdir (a copy of the simconf$workdir), C_emb (a SingleCellExperiment instance with cell embedding), G_emb (a SingleCellExperiment instance with gene embedding), pp_CG (a SingleCellExperiment instance with preprocessed representation of the input data).

Note

simba module is imported with convert=FALSE to help zellkonverter. In the example gout$gdf is a python reference that could go stale.

Examples

h5adpath = system.file(file.path("h5ad", "tenx3k.h5ad"), package="BiocSIMBA")
gout = gen_graph_sce(h5adpath, simconf = simba_config(gen_graph_copy=TRUE)) # generally want copy to be FALSE
#> Registered S3 methods overwritten by 'zellkonverter':
#>   method                                             from      
#>   py_to_r.numpy.ndarray                              reticulate
#>   py_to_r.pandas.core.arrays.categorical.Categorical reticulate
gout$gdf
#>          source relation destination
#> 0           C.1       r0        G.67
#> 1           C.1       r0        G.81
#> 2           C.1       r0       G.113
#> 3           C.1       r0       G.120
#> 4           C.1       r0       G.189
#> ...         ...      ...         ...
#> 2286879  C.2699       r4     G.30970
#> 2286880  C.2699       r4     G.31188
#> 2286881  C.2699       r4     G.31257
#> 2286882  C.2699       r4     G.31398
#> 2286883  C.2699       r4     G.32022
#> 
#> [2286884 rows x 3 columns]
dir(gout$work_contents, full.names=TRUE, recursive=TRUE)
#>  [1] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/entity_alias.txt"                  
#>  [2] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/graph_stats.json"                  
#>  [3] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/input/edge/edges_0_0.h5"           
#>  [4] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/input/entity/entity_count_C_0.txt" 
#>  [5] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/input/entity/entity_count_G_0.txt" 
#>  [6] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/input/entity/entity_names_C_0.json"
#>  [7] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/input/entity/entity_names_G_0.json"
#>  [8] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/model/checkpoint_version.txt"      
#>  [9] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/model/config.json"                 
#> [10] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/model/embeddings_C_0.v10.h5"       
#> [11] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/model/embeddings_G_0.v10.h5"       
#> [12] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/model/model.v10.h5"                
#> [13] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/model/training_stats.json"         
#> [14] "/home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work/pbg/graph0/pbg_graph.txt"                     
gout
#> simba_scrna instance based on input with 32738 genes and 2700 cells.
#>   The workdir used was /home/vincent/TEMP/RtmppfLAHp/file3245040ac210a/simba_work.
gout = annotate_C_emb(gout, "celltype")
names(colData(gout$C_emb))
#> [1] "celltype"